chore: refactoring to reduce the necessary file changes
