+++ callouts/cpp/examples/basic/BUILD
@@ -3,26 +3,27 @@ package(default_visibility = ["//visibility:public"])
 licenses(["notice"])  # Apache 2
 
 cc_binary(
-    name = "basic_cpp",
-    srcs = ["main.cc"],
+    name = "custom_callout_server_cpp",
+    srcs = ["//:service/main.cc"],
     deps = [
-        ":basic_cpp_lib",
+        ":custom_callout_server_lib",
     ],
+    includes = ["."],
 )
 
 cc_library(
-    name = "basic_cpp_lib",
-    srcs = ["basic.h"],
+    name = "custom_callout_server_lib",
+    srcs = ["custom_callout_server.h"],
     deps = [
         "//:callout_service_cpp_lib",
     ],
 )
 
 cc_test(
-    name = "basic_cpp_tests",
-    srcs = ["basic_test.cc"],
+    name = "custom_callout_server_tests",
+    srcs = ["custom_callout_server_test.cc"],
     deps = [
-        ":basic_cpp_lib",
+        ":custom_callout_server_lib",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
     ],
